pyproject.toml
[tool.poetry]
name = "repo_prompt"
version = "0.1.0"
description = "Tool to convert a Git repo to a prompt to be used with Generative AI models."
authors = ["Will Parr"]
license = "MIT"
readme = "README.md"
packages = [{include = "repo_to_prompt"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
click = "^8.1.7"
gitpython = "^3.1.43"
pathspec = "^0.12.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
pytest = "^8.3.2"
pre-commit = "^3.8.0"
mypy = "^1.14.1"
types-click = "^7.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
ignore = ["E266", "W503", "E203", "E402"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.isort]
profile = "black"
skip = [".yml"]

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = "."

[tool.poetry.scripts]
r2p = "repo_to_prompt.cli:main"
